# backend/api/rest/v1/swagger.yaml
definitions:
...
  github_com_JMURv_sso_internal_auth.Claims:
    properties:
      aud:
        description: the `aud` (Audience) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.3
        items:
          type: string
        type: array
      exp:
        allOf:
        - $ref: '#/definitions/jwt.NumericDate'
        description: the `exp` (Expiration Time) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.4
      iat:
        allOf:
        - $ref: '#/definitions/jwt.NumericDate'
        description: the `iat` (Issued At) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.6
      iss:
        description: the `iss` (Issuer) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.1
        type: string
      jti:
        description: the `jti` (JWT ID) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.7
        type: string
      nbf:
        allOf:
        - $ref: '#/definitions/jwt.NumericDate'
        description: the `nbf` (Not Before) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.5
      roles:
        items:
          $ref: '#/definitions/github_com_JMURv_sso_internal_models.Role'
        type: array
      sub:
        description: the `sub` (Subject) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.2
        type: string
      uid:
        type: string
    type: object
...
  github_com_JMURv_sso_internal_dto.CheckForgotPasswordEmailRequest:
    properties:
      password:
        type: string
      token:
        type: integer
      uidb64:
        type: string
    required:
    - password
    - token
    - uidb64
    type: object
  github_com_JMURv_sso_internal_dto.CheckLoginCodeRequest:
    properties:
      code:
        type: integer
      email:
        type: string
    required:
    - code
    - email
    type: object
...
  github_com_JMURv_sso_internal_dto.EmailAndPasswordRequest:
    properties:
      email:
        type: string
      password:
        type: string
      token:
        type: string
    required:
    - email
    - password
    - token
    type: object
...
  github_com_JMURv_sso_internal_dto.LoginCodeRequest:
    properties:
      email:
        type: string
      password:
        type: string
      token:
        type: string
    required:
    - email
    - password
    - token
    type: object
...
  github_com_JMURv_sso_internal_dto.RefreshRequest:
    properties:
      refresh:
        type: string
    required:
    - refresh
    type: object
  github_com_JMURv_sso_internal_dto.SendForgotPasswordEmail:
    properties:
      email:
        type: string
      token:
        type: string
    required:
    - email
    - token
    type: object
  github_com_JMURv_sso_internal_dto.TokenPair:
    properties:
      access:
        type: string
      refresh:
        type: string
    type: object
  github_com_JMURv_sso_internal_dto.TokenRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse:
    properties:
      errors:
        items:
          type: string
        type: array
    type: object
...
  github_com_JMURv_sso_internal_models.Permission:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_JMURv_sso_internal_models.Role:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_JMURv_sso_internal_models.Permission'
        type: array
    type: object
...
  jwt.NumericDate:
    properties:
      time.Time:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /auth/email/check:
    post:
      consumes:
      - application/json
      description: Exchange a valid email code for JWT tokens
      parameters:
      - description: Client real IP address
        in: header
        name: X-Real-IP
        required: true
        type: string
      - description: Client User-Agent
        in: header
        name: User-Agent
        required: true
        type: string
      - description: code, reCAPTCHA token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_sso_internal_dto.CheckLoginCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_dto.TokenPair'
        "400":
          description: missing device info or bad payload
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "404":
          description: code not found
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Check email login code
      tags:
      - EmailAuth
  /auth/email/send:
    post:
      consumes:
      - application/json
      description: Verify reCAPTCHA, then send a one-time code to the user’s email.
        May return tokens if password also valid.
      parameters:
      - description: Client real IP address
        in: header
        name: X-Real-IP
        required: true
        type: string
      - description: Client User-Agent
        in: header
        name: User-Agent
        required: true
        type: string
      - description: email, password, reCAPTCHA token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_sso_internal_dto.LoginCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_dto.TokenPair'
        "400":
          description: missing device info or bad payload
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "401":
          description: invalid credentials or reCAPTCHA
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Send login code via email
      tags:
      - EmailAuth
  /auth/jwt:
    post:
      consumes:
      - application/json
      description: Verify reCAPTCHA, then authenticate and set JWT cookies
      parameters:
      - description: Client real IP address
        in: header
        name: X-Real-IP
        required: true
        type: string
      - description: Client User-Agent
        in: header
        name: User-Agent
        required: true
        type: string
      - description: email, password, reCAPTCHA token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_sso_internal_dto.EmailAndPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_dto.TokenPair'
        "400":
          description: missing device info or bad payload
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "401":
          description: invalid credentials or reCAPTCHA
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Authenticate using email & password
      tags:
      - Authentication
  /auth/jwt/parse:
    post:
      consumes:
      - application/json
      description: Decode a token without requiring device header
      parameters:
      - description: jwt token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_sso_internal_dto.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_auth.Claims'
        "404":
          description: token not found
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Parse JWT claims
      tags:
      - Authentication
  /auth/jwt/refresh:
    post:
      consumes:
      - application/json
      description: Validate device header and refresh tokens, reset cookies
      parameters:
      - description: Client real IP address
        in: header
        name: X-Real-IP
        required: true
        type: string
      - description: Client User-Agent
        in: header
        name: User-Agent
        required: true
        type: string
      - description: refresh_token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_sso_internal_dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_dto.TokenPair'
        "400":
          description: missing device info or bad payload
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "401":
          description: token revoked or invalid
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "404":
          description: session not found
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Refresh JWT tokens
      tags:
      - Authentication
  /auth/logout:
    post:
      description: Revoke refresh token, clear JWT cookies
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: session not found
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Logout user
      tags:
      - Authentication
...
  /auth/recovery/check:
    post:
      consumes:
      - application/json
      description: Validate a recovery code
      parameters:
      - description: code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_sso_internal_dto.CheckForgotPasswordEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: invalid code
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "404":
          description: code not found
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Check forgot‐password code
      tags:
      - PasswordRecovery
  /auth/recovery/send:
    post:
      consumes:
      - application/json
      description: Verify reCAPTCHA and send recovery email
      parameters:
      - description: email, reCAPTCHA token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_sso_internal_dto.SendForgotPasswordEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: email not found
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_JMURv_sso_internal_hdl_http_utils.ErrorsResponse'
      summary: Send forgot‐password email
      tags:
      - PasswordRecovery
...
swagger: "2.0"